create network and layers

+++ src/math/vector.rs
@@ -1,7 +1,7 @@
 use rand::Rng;
 use rand_distr::{Distribution, Normal};
 
-pub fn create_random_vector(size: usize) -> Vec<f64> {
+pub fn create_random_vector(size: u16) -> Vec<f64> {
     let mut rng = rand::thread_rng();
     let distribution = Normal::new(0.0, 1.0).unwrap();
     let mut vector: Vec<f64> = Vec::new();
